separate collectible-card/Footer  into components to be able to manage it well
