+++ sdk/src/react/ui/components/marketplace-collectible-card/components/footer/index.ts
@@ -1,0 +1,2 @@
+export * from './components';
+export * from './Footer.tsx';
